requirements.txt
absl-py==1.4.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
blis==0.7.10
cachetools==5.3.1
capstone==5.0.0
catalogue==2.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
clip==0.2.0
colored-traceback==0.3.0
comm==0.1.2
conda-pack==0.6.0
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
Cython==3.0.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
docopt==0.6.2
executing==1.2.0
fairscale==0.4.4
fastjsonschema==2.16.3
filelock==3.9.0
fonttools==4.42.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.6.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h5py==3.9.0
huggingface-hub==0.16.4
idna==3.4
intervaltree==3.1.0
ipykernel==6.21.3
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.4.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langcodes==3.3.0
lightning-utilities==0.10.1
Mako==1.2.4
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
mistune==2.0.5
multidict==6.0.5
murmurhash==1.0.9
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.16.1
nbformat==5.7.3
nest-asyncio==1.5.6
ninja==1.11.1.1
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.0
pandas==2.0.3
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.0.0
plumbum==1.8.1
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pwn==1.0
pwntools==4.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pyelftools==0.29
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==2.2.0.post0
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
ROPGadget==7.3
rpyc==5.3.1
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
safetensors==0.3.1
scipy==1.11.3
Send2Trash==1.8.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
stack-data==0.6.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
terminado==0.17.1
thinc==8.1.12
timm==0.4.12
tinycss2==1.2.1
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchmetrics==1.3.1
torchtext==0.4.0
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
unicorn==2.0.1.post1
uri-template==1.2.0
urllib3==1.26.16
virtualenv==20.19.0
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.6
widgetsnbextension==4.0.5
yacs==0.1.8
yarg==0.1.9
yarl==1.9.4
